Move config to core
