+++ src/main/java/io/javalin/core/JavalinServer.kt
@@ -7,7 +7,6 @@
 package io.javalin.core
 
 import io.javalin.Javalin
-import io.javalin.JavalinConfig
 import io.javalin.http.JavalinServlet
 import io.javalin.websocket.JavalinWsServlet
 import io.javalin.websocket.isWebSocket
