+++ src/main/java/io/javalin/http/JavalinServlet.kt
@@ -6,7 +6,7 @@
 
 package io.javalin.http
 
-import io.javalin.JavalinConfig
+import io.javalin.core.JavalinConfig
 import io.javalin.core.CachedRequestWrapper
 import io.javalin.core.security.CoreRoles
 import io.javalin.core.security.Role
